refactor(identity): adapter 응답 import 경로 정리 #48
