+++ systems/identity/identity-adapter-in/src/main/kotlin/hs/kr/entrydsm/identity/adapterin/web/dto/response/ApplicationStatusResponse.kt
@@ -1,6 +1,6 @@
 package hs.kr.entrydsm.identity.adapterin.web.dto.response
 
-import hs.kr.entrydsm.identity.domain.ApplicantStatus
+import hs.kr.entrydsm.identity.domain.enum.ApplicantStatus
 import java.time.Instant
 
 data class ApplicationStatusResponse(
